Rename base client module to 'client'
`globus_sdk.client` is a rename of `globus_sdk.base`
