+++ tests/functional/base_client/conftest.py
@@ -5,7 +5,7 @@ import globus_sdk
 
 @pytest.fixture
 def client():
-    class CustomClient(globus_sdk.base.BaseClient):
+    class CustomClient(globus_sdk.client.BaseClient):
         service_name = "foo"
 
     return CustomClient()
